requirements.txt
google-generativeai==0.7.2
python-dotenv==1.0.1
streamlit==1.39.0
streamlit-option-menu==0.3.12
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.0
nltk==3.9.1
textblob==0.17.1
langdetect==1.0.9
textstat==0.7.3
googletrans==4.0.0rc1
wordcloud==1.9.3
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0
tiktoken==0.7.0
transformers==4.44.2
torch==2.5.1
pyarrow==17.0.0
jinja2==3.1.4
gradio
panel
fastapi
uvicorn
python-multipart

# --- ADD THESE FOR FINE-TUNING ---
datasets
peft
accelerate
bitsandbytes
trl
scipy

# --- ADD THESE FOR AI AGENTS & RAG ---
chromadb==0.4.22
onnxruntime==1.16.3
faiss-cpu==1.7.4
google-api-python-client==2.149.0
pytrends==4.9.2
opencv-python==4.10.0.84
language-tool-python==2.7.1
